requirements.txt
web3==5.31.1
python-dateutil==2.7.0
click==8.0.1
ethereum-dasm==0.1.4
base58
requests
pytest~=4.3.0
timeout-decorator==0.4.1
sqlalchemy==1.3.13
pymongo==3.11.4
pyetherchain==0.3.2
QueryStateLib==1.1.4
prometheus-client
python-dotenv